Update Gene Annotations from Ensembl

Retrieves current gene annotations for probe genomic locations using biomaRt. Maps probes to genes based on genomic overlap.

update_gene_annotations(
  probe_ranges,
  species = "hsapiens",
  ensembl_version = NULL,
  host = NULL,
  gene_biotypes = NULL,
  expand_range = 0
)

Arguments

probe_ranges

GRanges object with probe locations (from align_probes_to_genome)

species

Species name: "hsapiens", "mmusculus", "rnorvegicus", etc. (default: "hsapiens")

ensembl_version

Ensembl version to use (default: NULL for current)

host

Ensembl host URL (default: NULL for main site)

gene_biotypes

Gene biotypes to include (default: NULL for all)

expand_range

Expand probe range by N bp for nearby genes (default: 0)

Value

Data frame with updated gene annotations

Examples

if (FALSE) { # \dontrun{
  # Update annotations for human probes
  annotations <- update_gene_annotations(
    probe_ranges,
    species = "hsapiens"
  )

  # Include nearby genes within 5kb
  annotations <- update_gene_annotations(
    probe_ranges,
    species = "hsapiens",
    expand_range = 5000
  )
} # }